[refactor] #93 HomeBinding HomeViewModel
