+++ app/src/main/java/com/milkyway/milkyway/ui/home/HomeBinding.kt
@@ -1,6 +1,7 @@
 package com.milkyway.milkyway.ui.home
 
 import android.widget.ImageButton
+import android.widget.TextView
 import androidx.databinding.BindingAdapter
 import com.milkyway.milkyway.R
 
@@ -10,5 +11,11 @@ object HomeBinding {
     fun setButtonChange(button : ImageButton, compass : Boolean) {
         if(compass) button.setBackgroundResource(R.drawable.btn_compass)
         else button.setBackgroundResource(R.drawable.btn_current_location)
+    }
+
+    @BindingAdapter("app:isSelected")
+    @JvmStatic
+    fun isSelected(textView : TextView, status: Boolean) {
+        textView.isSelected = status
     }
 }
