--- app/src/main/java/com/milkyway/milkyway/ui/home/bindingAdapter.kt
@@ -1,9 +1,0 @@
-package com.milkyway.milkyway.ui.home
-
-import android.view.View
-import androidx.databinding.BindingAdapter
-
-@BindingAdapter("app:isSelected")
-fun View.isSelected(status: Boolean) {
-    isSelected = status
-}
